Add BottomSheet and change logic database for Playlist

+++ app/src/main/java/com/example/playlistmaker/media/domain/db/FavoriteTracksInteractor.kt
@@ -11,7 +11,7 @@ interface FavoriteTracksInteractor {
 
     fun getFavoriteTracks(): Flow<List<Track>>
 
-    suspend fun isTrackFavorite(trackId: Int): Boolean
+    suspend fun isTrackFavorite(trackId: Long): Boolean
 }
